+++ app/src/main/java/org/cxct/sportlottery/ui/maintab/games/adapter/RecyclerLiveListAdapter.kt
@@ -36,7 +36,7 @@ class RecyclerLiveListAdapter : BindingAdapter<OKGamesCategory, ItemLiveViewList
 
                 var changed = false
                 okGamesCategory.gameList?.forEachIndexed { position, okGameBean->
-                    if (okGameBean.isMaintain() != (gamesMaintain.maintain.toInt() == 1) && (okGameBean.firmType == gamesMaintain.firmType)) {
+                    if (okGameBean.isMaintain() != (gamesMaintain.maintain == 1) && (okGameBean.firmType == gamesMaintain.firmType)) {
                         okGameBean.maintain = gamesMaintain.maintain.toInt()
                         changed = true
                     }
