+++ app/src/main/java/org/cxct/sportlottery/ui/maintab/games/view/RecyclerGamePageAdapter.kt
@@ -63,7 +63,7 @@ class RecyclerGamePageAdapter:
                 val changedPosition = mutableListOf<Pair<Int, OKGameBean>>()
 
                 okGameBeans.forEachIndexed { position, gameBean->
-                    if (gameBean.isMaintain() != (gamesMaintain.maintain.toInt() == 1) && (gameBean.firmType == gamesMaintain.firmType)) {
+                    if (gameBean.isMaintain() != (gamesMaintain.maintain == 1) && (gameBean.firmType == gamesMaintain.firmType)) {
                         gameBean.maintain = gamesMaintain.maintain.toInt()
                         changedPosition.add(Pair(position, gameBean))
                     }
